Refactor tests, move _parseDid().
