+++ package.json
@@ -78,7 +78,7 @@
   "scripts": {
     "test": "npm run test-node && npm run test-karma",
     "test-node": "cross-env NODE_ENV=test mocha -r esm --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} test/**/*.spec.js",
-    "test-karma": "karma start test/karma.conf.js",
+    "test-karma": "karma start karma.conf.js",
     "lint": "eslint .",
     "preversion": "npm test"
   },
